requirement.txt
torch>=1.6
torchvision
numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb
kornia=0.6.2
pyciede2000
pyiqa==0.1.5
